Added db constants and reorganized it a bit

+++ db.lua
@@ -1,49 +1,51 @@
 module(..., package.seeall)
 
 require "sqlite3"
---
----- deblare the class table first
---local db = {}
---
+
 ---- local variables and constants
-local database
+local database = nil
+-- database query strings
+local data_SelectAllFromItem = [[ Select * from item; ]]
+
+function checkDataBaseState()
+    if (database == nil or database.isopen(database) == false) then
+        local path = system.pathForFile("data.db", system.DocumentsDirectory)
+        database = sqlite3.open(path)
+    end
+end
 
 function printStuff()
-    local t
-    for row in database:nrows("SELECT * FROM item") do
-            print("found items")
-      local t = display.newText(row.name, 20, 30 * row.id, null, 16)
-      t:setTextColor(255,0,255)
+    checkDataBaseState{}
+    for row in database:nrows(data_SelectAllFromItem) do
+        local t = display.newText(row.name, 20, 30 * row.id, null, 16)
+        t:setTextColor(255,0,255)
     end    
-    return t
 end
 
 function initialize()
-    local path = system.pathForFile("data.db", system.DocumentsDirectory)
-    database = sqlite3.open(path)
+    checkDataBaseState{}
     
-    print("before execute")
-    --function database:initialize()
-    database:exec[[
-            PRAGMA foreign_keys = ON;
+    rows = database:nrows([[ SELECT * FROM item ]])
+    
+    if (rows ~= "0") then
+        database:execute(
+        [[ PRAGMA foreign_keys = ON;
             drop table if exists purchase;
             drop table if exists item;
             CREATE TABLE item (id integer PRIMARY KEY, name text, active integer);
             CREATE TABLE purchase (id integer PRIMARY KEY, itemid integer,
                     date integer, amount integer, active integer,
-                    FOREIGN KEY(itemid) REFERENCES item(id));
-    ]]
+                    FOREIGN KEY(itemid) REFERENCES item(id)); 
+        ]])
+    end
     
-    
-    database:exec[[
-            insert into item (id, name, active) values (null, 'Diapers', 1);
-            insert into item (id, name, active) values (null, 'Prius Gas', 1);
-            insert into purchase (id, itemid, date, amount, active) values (null, 2, DATETIME('now'), 33.30, 1);
-    ]]
-    --end
-    print("after execute")
-    
-    print( "version " .. sqlite3.version() )
+    database:execute(
+    [[
+        insert into item (id, name, active) values (null, 'Diapers', 1);
+        insert into item (id, name, active) values (null, 'Prius Gas', 1);
+        insert into purchase (id, itemid, date, amount, active) values (null, 2, DATETIME('now'), 33.30, 1);
+    ]])
+        
     return database
 end
 --
